requirements.txt.txt
requests
pygame
mixer
from io import BytesIO
pillow
difflib
unittest